Made my test a little harder
Adding my first attempt. It worked on my test. I am running it on the challenge now. It is really slow and would benefit from multi-threading, but I have to call it quits. At least for now.

Still need to add the part where it prints the output to a file, which is simple.